refactor(cli): move argument handling out of main
