+++ tests/cli.rs
@@ -1,21 +1,7 @@
-use std::process::Command;
-use std::{
-    fs,
-    path::PathBuf,
-    time::{SystemTime, UNIX_EPOCH},
-};
+mod support;
 
-fn elio() -> Command {
-    Command::new(env!("CARGO_BIN_EXE_elio"))
-}
-
-fn temp_path(label: &str) -> PathBuf {
-    let unique = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("system time should be after unix epoch")
-        .as_nanos();
-    std::env::temp_dir().join(format!("elio-cli-{label}-{unique}"))
-}
+use std::fs;
+use support::{elio, temp_path};
 
 #[test]
 fn version_prints_package_version() {
